\name{bma.expand}
\alias{bma.expand}
\title{bma.expand}
\usage{
  bma.expand(data, bma, groups)
}
\arguments{
  \item{data}{object of class snpBMAdata}

  \item{bma}{object of class snpBMA}

  \item{groups}{named list, with names equal to all SNPs to
  be included and}
}
\value{
  object of class snpBMA
}
\description{
  Expand SNP models to include selected tagged SNPs
}
\author{
  Chris Wallace
}

