\name{bma.grow}
\alias{bma.grow}
\title{bma.grow}
\usage{
  bma.grow(data, bma)
}
\arguments{
  \item{data}{object of class snpBMAdata}

  \item{bma}{object of class snpBMA}
}
\value{
  object of class snpBMA
}
\description{
  Grow the next generation of SNP models from the current
  one
}
\author{
  Chris Wallace
}

