\name{mcomp.detail}
\alias{mcomp.detail}
\title{mcomp.detail}
\usage{
  mcomp.detail(m.parent, m.child, bf.parent, bf.child,
    ntarget, n.child, prior.parents, prior.children, what,
    pp.fold = 10, quiet = FALSE)
}
\arguments{
  \item{m.parent}{}

  \item{m.child}{}

  \item{bf.parent}{}

  \item{bf.child}{}

  \item{ntarget}{}

  \item{n.child}{}

  \item{prior.parents}{}

  \item{prior.children}{}

  \item{what}{}

  \item{pp.fold}{}

  \item{quiet}{}
}
\value{
  object of class dropModels, or an index vector of which
  rows of supplied m.child should be dropped
}
\description{
  mcomp.detail, internal function
}
\author{
  Chris Wallace
}

