\name{my.glib}
\alias{my.glib}
\title{my.glib}
\usage{
  my.glib(data, models)
}
\arguments{
  \item{data}{snpBMAdata object}

  \item{models}{matrix of models to fit}
}
\value{
  list containing the essential results from glib
}
\description{
  Internal function, wrapper for BMA's glib
}
\author{
  Chris Wallace
}

