\docType{methods}
\name{show}
\alias{show}
\alias{show,snpBMA-method}
\alias{show,snpBMAdata-method}
\alias{show,snpBMAlist-method}
\title{show-methods}
\arguments{
  \item{object}{object of class snpBMA, snpBMAdata or
  snpBMAlist}
}
\value{
  no return value
}
\description{
  Show snpBMA and related objects
}
\details{
  For all objects, a very brief description of the object
  size is given.  For snpBMA objects only, the models with
  greatest support are shown.
}
\author{
  Chris Wallace
}
\keyword{methods}

