\docType{methods}
\name{snps.prune}
\alias{snps.prune}
\alias{snps.prune,snpBMA-method}
\alias{snps.prune,snpBMAdata-method}
\alias{snps.prune,snpBMAstrat-method}
\alias{snps.prune-methods}
\title{snps.prune}
\arguments{
  \item{object}{a snpBMAdata or snpBMAstrat object}

  \item{prune.snps}{character vector of SNPs to drop}

  \item{...}{other arguments, currently ignored}
}
\value{
  object of the same class as \code{object}
}
\description{
  Drop SNPs from a snpBMAdata or snpBMAstrat object
}
\author{
  Chris Wallace
}
\keyword{methods}

