\docType{methods}
\name{stack}
\alias{stack}
\alias{stack,modelSummary-method}
\alias{stack,snpBMA-method}
\alias{stack-methods}
\title{stack-methods}
\arguments{
  \item{object}{object of class snpBMA or modelSummary}

  \item{...}{more objects of the same class}
}
\value{
  object of class snpBMAlist or modelSummary
}
\description{
  Stack objects together
}
\author{
  Chris Wallace
}
\keyword{methods}

