\name{top.snps}
\alias{top.snps}
\title{top.snps}
\usage{
  top.snps(object, snpsep = "/", ...)
}
\arguments{
  \item{object}{snpBMA or snpBMAlist object}

  \item{...}{additional arguments to pass to top.models,
  notably \code{nmodels}: the number of models to include}
}
\value{
  character vector of SNP names
}
\description{
  Extract SNPs from top.models
}
\author{
  Chris Wallace
}

