\docType{methods}
\name{stack}
\alias{stack}
\alias{stack,list-method}
\alias{stack,snpBMA-method}
\alias{stack-methods}
\title{stack-methods}
\arguments{
  \item{object}{object of class snpBMA}

  \item{...}{more objects of class snpBMA}
}
\value{
  object of class snpBMAlist
}
\description{
  Stack snpBMA objects together
}
\author{
  Chris Wallace
}
\keyword{methods}

