\docType{methods}
\name{top.models}
\alias{top.models}
\alias{top.models,snpBMA-method}
\alias{top.models,snpBMAlist-method}
\alias{top.models-methods}
\title{top.models}
\arguments{
  \item{object}{object of class snpBMA or snpBMAlist}

  \item{...}{number of models to show}
}
\value{
  no return value, models are printed to screen
}
\description{
  Show models with greatest posterior probability
}
\author{
  Chris Wallace
}
\keyword{methods}

