\name{top.snpBMAlist}
\alias{top.snpBMAlist}
\title{snpBMA}
\usage{
  top.snpBMAlist(object, priors, nmodels = 6L)
}
\arguments{
  \item{object}{Object of class snpBMA}

  \item{nmodels}{number of models to show}
}
\value{
  matrix of top models and their Bayes Factors
}
\description{
  Show top models in a snpBMA object
}
\author{
  Chris Wallace
}

